+++ packages/typography/src/Description/Description.styles.ts
@@ -1,0 +1,40 @@
+import { css } from '@leafygreen-ui/emotion';
+import { Theme } from '@leafygreen-ui/lib';
+import { BaseFontSize, fontFamilies, typeScales } from '@leafygreen-ui/tokens';
+import { palette } from '@leafygreen-ui/palette';
+
+export const descriptionStyle = css`
+  font-family: ${fontFamilies.default};
+  font-weight: normal;
+  margin-top: 0;
+  margin-bottom: 0;
+`;
+
+export const descriptionColorStyle: Record<Theme, string> = {
+  [Theme.Light]: css`
+    color: ${palette.gray.dark1};
+  `,
+  [Theme.Dark]: css`
+    color: ${palette.gray.light1};
+  `,
+};
+
+export const disabledDescriptionColorStyle: Record<Theme, string> = {
+  [Theme.Light]: css`
+    color: ${palette.gray.dark1};
+  `,
+  [Theme.Dark]: css`
+    color: ${palette.gray.base};
+  `,
+};
+
+export const descriptionTypeScaleStyles: Record<BaseFontSize, string> = {
+  [BaseFontSize.Body1]: css`
+    font-size: ${typeScales.body1.fontSize}px;
+    line-height: ${typeScales.body1.lineHeight}px;
+  `,
+  [BaseFontSize.Body2]: css`
+    font-size: ${typeScales.body2.fontSize}px;
+    line-height: 20px; // Hardcoding because it does not match body2 lineHeight
+  `,
+};