+++ packages/typography/src/H3/H3.types.ts
@@ -1,0 +1,4 @@
+import { HTMLElementProps } from '@leafygreen-ui/lib';
+import { CommonTypographyProps } from '../types';
+
+export type H3Props = HTMLElementProps<'h3'> & CommonTypographyProps;