+++ packages/menu/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @leafygreen-ui/menu
 
+## 19.0.1
+
+### Patch Changes
+
+- 1a335d0b2: Migrate component internals to check for glyphs explicity, rather than for Icon components as well
+- ae5421cf6: Updates components to use internal transition tokens
+- Updated dependencies [ae5421cf6]
+- Updated dependencies [4b4c2d27d]
+- Updated dependencies [1a335d0b2]
+- Updated dependencies [ae5421cf6]
+  - @leafygreen-ui/tokens@1.4.0
+  - @leafygreen-ui/icon@11.12.1
+  - @leafygreen-ui/icon-button@15.0.1
+  - @leafygreen-ui/popover@11.0.1
+
 ## 19.0.0
 
 ### Patch Changes