+++ packages/pipeline/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @leafygreen-ui/pipeline
 
+## 5.0.1
+
+### Patch Changes
+
+- ae5421cf6: Updates components to use internal transition tokens
+- Updated dependencies [ae5421cf6]
+- Updated dependencies [ed0e425e5]
+- Updated dependencies [4b4c2d27d]
+- Updated dependencies [1a335d0b2]
+  - @leafygreen-ui/tokens@1.4.0
+  - @leafygreen-ui/tooltip@9.0.1
+  - @leafygreen-ui/icon@11.12.1
+
 ## 5.0.0
 
 ### Patch Changes